+++ clang/test/SemaTemplate/cwg2398.cpp
@@ -94,12 +94,10 @@ namespace type_pack3 {
   template<class T3> struct B;
 
   template<template<class T4              > class TT1, class T5              > struct B<TT1<T5        >>;
-  // new-note@-1 {{template is declared here}}
-  template<template<class T6, class ...T7s> class TT2, class T8, class ...T9s> struct B<TT2<T8, T9s...>>;
-  // old-note@-1 {{template is declared here}}
+
+  template<template<class T6, class ...T7s> class TT2, class T8, class ...T9s> struct B<TT2<T8, T9s...>> {};
 
   template struct B<A<int>>;
-  // expected-error@-1 {{explicit instantiation of undefined template}}
 } // namespace type_pack3
 
 namespace gcc_issue {
@@ -301,12 +299,14 @@ namespace classes {
     };
 
     template <template <class T3> class TT> void f(TT<int> v) {
-      // old-note@-1 {{template template argument has different template parameters}}
-      // new-note@-2 {{deduced type 'A<[...], (no argument), (no argument), (no argument)>' of 1st parameter does not match adjusted type 'A<[...], void, void, void>' of argument [with TT = A:1<void, void, void>]}}
+      // new-note@-1 {{substitution failure: too many template parameters in template template parameter}}
+      // old-note@-2 {{template template argument has different template parameters}}
       static_assert(v.val == 3);
     };
     void test() {
-      // FIXME: Needs deduction of defaulted template parameter packs.
+      // FIXME: We incorrectly reject the pack to non-pack matching.
+      // This should apply the same rules as directly specifying the arguments,
+      // but the partial ordering rules are applied instead.
       f(A<int, void, void, void>());
       // expected-error@-1 {{no matching function for call}}
     }
@@ -351,6 +351,19 @@ namespace classes {
     // old-error@-1 {{no matching function for call}}
   } // namespace defaulted
 } // namespace classes
+
+namespace packs {
+  namespace t1 {
+    template<template<class, int, int...> class> struct A {};
+    // old-note@-1 {{previous non-type template parameter with type 'int' is here}}
+
+    template<class, char> struct B;
+    // old-note@-1 {{template non-type parameter has a different type 'char' in template argument}}
+
+    template struct A<B>;
+    // old-error@-1 {{has different template parameters}}
+  } // namespace t1
+} // namespace packs
 
 namespace regression1 {
   template <typename T, typename Y> struct map {};